# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "name": "Your Name",
            "email": "you@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Get all orders",
                "description": "Endpoint to retrieve a list of all orders.",
                "operationId": "6e2e0d5c90ed4a866d1965de8bca6f24",
                "responses": {
                    "200": {
                        "description": "List of orders"
                    }
                }
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Create a new orders",
                "description": "Endpoint to create a new orders.",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "The ID of the user for whom the order is placed",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "The title of the order",
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "description": "The description of the order",
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "description": "The creation date and time of the order",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "updated_at": {
                                        "description": "The last update date and time of the order",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "order created successfully"
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    }
                }
            }
        },
        "/api/orders/{order_id}": {
            "put": {
                "tags": [
                    "orders"
                ],
                "summary": "Update order details",
                "description": "Endpoint to update order details.",
                "operationId": "d3243d77d764ebfd6d8ae657f386d468",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "ID of the order to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "The ID of the user for whom the order is placed",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "The title of the order",
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "description": "The description of the order",
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "description": "The creation date and time of the order",
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "updated_at": {
                                        "description": "The last update date and time of the order",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification details updated successfully"
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "orders"
                ],
                "summary": "Delete a order",
                "description": "Endpoint to delete a notification by ID.",
                "operationId": "c05cd1972f9fba85cff8f18a2a1d8567",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "ID of the order to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification deleted successfully"
                    },
                    "404": {
                        "description": "Notification not found"
                    }
                }
            }
        },
...
    }
}